Extend node APIs: - Implement GetNodeState - Implement KillAllNodesJobs - Implement KillNodeJob - Modify CPUInfo properties - Modify CoreStateResult properties

+++ CoreStateResult.go
@@ -6,10 +6,10 @@ import (
 
 // Result of a core.state command
 type CoreStateResult struct {
-	Cpu  int `json:"cpu" validate:"nonzero"`
-	Rss  int `json:"rss" validate:"nonzero"`
-	Swap int `json:"swap" validate:"nonzero"`
-	Vms  int `json:"vms" validate:"nonzero"`
+	Cpu  float64 `json:"cpu" validate:"nonzero"`
+	Rss  int64   `json:"rss" validate:"nonzero"`
+	Swap int64   `json:"swap" validate:"nonzero"`
+	Vms  int64   `json:"vms" validate:"nonzero"`
 }
 
 func (s CoreStateResult) Validate() error {
